Clean up code by introducing async/await
This style makes the code somewhat easier to read.  We shifted away
from async/await because mollie did not have it in their documentation
and we ran into some issues.  We suspected the mollie api might not
have been real promises.  Looks like this adaptation works the way we
want it to.