Write status based on webhook
When we receive the status from the Mollie API, we now set it as a
string on the basket.  This should probably become a smarter semantic
model in the future, yet it should also suffice for now.

The status can be rendered in the frontend to indicate what happened
with the basket.  The construction with a trivial string can later be
extended so the backend talks in URIs whilst the frontend keeps its
current string properties.