Split difference between draft and completed baskets
There's now a difference between the draft baskets and the completed
baskets.  This means we can make a difference between them in the
payment page and that the user can order multiple baskets.

For the payment service that means we now indicate whether the basket
is still in draft or not, depending on the payment.  This might be
moved into a reactive service in the future in order to make the
payment service more reusable.